Remove modelset_config_test.go
This does not test code in provisiond (code coverage 2.6%), and the only test that isn't already in configd:session/component_config_test.go (the one for ordered lists) is better moved there anyway (done in separate commit). That avoids unnecessary rework of this test to work with changes being made for VRVDR-50443 in configd etc.

Part of VRVDR-50443.